A pack of changes... I started working to spruce up the code and when I finished, I had already made the following changes:
 - changed stopwatch.update function interface: it only receives the
Canvas parameter now;
 - moved clock positioning control variables as internal member class variables
 - added new member functions to reset the stopwatch counter, blanking area
when redrawing the clock numbers and finally the position of clock

Those changes were required for me to test the stopwatch in 3 different device
models probing the better parameter values for positioning and blanking area
(for redrawing the numbers).

Advantages of the changes: with the clock 'alive', you can change:
 - positioning
 - blanking area
 - toggle (pause/continue)
 - reset the counter.

TODO: add client GUI menu options to control the stopwatch.